less french words
